[Terry N.] code cleanups
git-svn-id: svn+ssh://eos.nps.edu/usr/local/svn/wcm/Viskit/trunk@2362 0de750ab-c0a9-dc11-9d74-0011435b89d6

+++ src/edu/nps/util/CryptoMethods.java
@@ -136,7 +136,7 @@ public class CryptoMethods
     StringBuilder sb = new StringBuilder();
     for (int i = 0; i < bytes.length; i++) {
       byte b = bytes[i];
-      sb.append((int) (0x00FF & b));
+      sb.append(0x00FF & b);
       if (i + 1 < bytes.length) {
         sb.append("-");
       }
